pyproject.toml
[project]
name = "libguard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.2",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-rest-framework>=0.1.0",
    "djangorestframework-simplejwt>=5.5.1",
    "google-auth>=2.48.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "rest-framework-simplejwt>=0.0.2",
]
